routes: migrate projects domain (simple GETs/CRUD, 13 routes)
Move to src/server/routes/projects.ts:
- POST /api/projects/detect, POST /api/projects/scan
- GET /api/projects/:id/structured, POST /api/projects/:id/rescan-repos
- GET /api/browse-directory
- GET /api/projects (list)
- GET /api/projects/:id, PUT, DELETE
- POST /api/projects/:id/promote
- GET /api/projects/:id/config, /defaults, /resolved
- GET /api/projects/:id/qa-testing-config

Dispatcher catches these BEFORE the legacy handleApiRoute. Two complex
handlers stay in server.ts for now:
- POST /api/projects (project creation with worktree-pool init)
- PUT /api/projects/:id/config (component validation, secret merging)

These are tightly coupled to project-context lifecycle and worth their own
follow-up commit.

Net: server.ts -255 lines (5664 → 5409). No behaviour change.
