+++ src/server/server.ts
@@ -1134,175 +1134,7 @@ async function handleApiRoute(
 
 	// ── Project Detection & Browse ────────────────────────────────────
 
-	// POST /api/projects/detect
-	if (url.pathname === "/api/projects/detect" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body || typeof body.path !== "string") {
-			json({ error: "Missing path" }, 400);
-			return;
-		}
-		const dirPath = path.resolve(body.path);
-		const exists = fs.existsSync(dirPath);
-		let hasBobbit = false;
-		let isEmpty = true;
-		let hasPackageJson = false;
-		let hasCargoToml = false;
-		let hasGoMod = false;
-		let name = path.basename(dirPath);
-
-		if (exists) {
-			try {
-				const stat = fs.statSync(dirPath);
-				if (stat.isDirectory()) {
-					const entries = fs.readdirSync(dirPath);
-					isEmpty = entries.length === 0;
-					hasBobbit = entries.includes(".bobbit");
-					hasPackageJson = entries.includes("package.json");
-					hasCargoToml = entries.includes("Cargo.toml");
-					hasGoMod = entries.includes("go.mod");
-
-					// Try to read name from package.json
-					if (hasPackageJson) {
-						try {
-							const pkg = JSON.parse(fs.readFileSync(path.join(dirPath, "package.json"), "utf-8"));
-							if (typeof pkg.name === "string" && pkg.name) {
-								name = pkg.name;
-							}
-						} catch {
-							// Ignore parse errors — fall back to directory basename
-						}
-					}
-				} else {
-					// Path exists but is not a directory
-					json({ error: "Path is not a directory" }, 400);
-					return;
-				}
-			} catch {
-				// stat failed — treat as non-existent
-				json({ exists: false, hasBobbit: false, isEmpty: true, hasPackageJson: false, hasCargoToml: false, hasGoMod: false, name });
-				return;
-			}
-		}
-
-		json({ exists, hasBobbit, isEmpty, hasPackageJson, hasCargoToml, hasGoMod, name });
-		return;
-	}
-
-	// POST /api/projects/scan?path=...  → run repo-scan on a folder.
-	// Returns { repos: DetectedRepo[] }. Used by the Add-Project flow and
-	// Settings → "Re-scan repos". Phase 4b — see docs/design/multi-repo-components.md §8.1.
-	if (url.pathname === "/api/projects/scan" && req.method === "POST") {
-		const body = await readBody(req).catch(() => ({}));
-		const rawPath = url.searchParams.get("path") ?? (body && typeof body.path === "string" ? body.path : "");
-		if (!rawPath) { json({ error: "Missing path" }, 400); return; }
-		const dirPath = path.resolve(rawPath);
-		if (!fs.existsSync(dirPath)) { json({ error: "Path not found" }, 404); return; }
-		try {
-			const { scanRepos } = await import("./agent/repo-scan.js");
-			const { scanMonorepo } = await import("./agent/monorepo-scan.js");
-			const repos = await scanRepos(dirPath);
-			const monorepo = scanMonorepo(dirPath);
-			json({ repos, monorepo });
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// GET /api/projects/:id/structured  → returns { components, workflows,
-	// worktree_root } in their structured (non-string) shape. Used by the
-	// Settings → Components tab so the UI doesn't have to parse YAML.
-	// Phase 4b.
-	const projectStructuredMatch = url.pathname.match(/^\/api\/projects\/([^/]+)\/structured$/);
-	if (projectStructuredMatch && req.method === "GET") {
-		const ctx = projectContextManager.getOrCreate(projectStructuredMatch[1]);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		const components = ctx.projectConfigStore.getComponents();
-		const workflows = ctx.projectConfigStore.getWorkflows() ?? {};
-		const worktreeRoot = ctx.projectConfigStore.get("worktree_root") ?? "";
-		json({ components, workflows, worktree_root: worktreeRoot });
-		return;
-	}
-
-	// POST /api/projects/:id/rescan-repos  → re-run repo-scan on the
-	// project's rootPath; returns the same shape as /api/projects/scan.
-	// Settings "Re-scan repos" button. Phase 4b.
-	const projectRescanMatch = url.pathname.match(/^\/api\/projects\/([^/]+)\/rescan-repos$/);
-	if (projectRescanMatch && req.method === "POST") {
-		const project = projectRegistry.get(projectRescanMatch[1]);
-		if (!project) { json({ error: "Project not found" }, 404); return; }
-		try {
-			const { scanRepos } = await import("./agent/repo-scan.js");
-			const { scanMonorepo } = await import("./agent/monorepo-scan.js");
-			const repos = await scanRepos(project.rootPath);
-			const monorepo = scanMonorepo(project.rootPath);
-			json({ repos, monorepo, rootPath: project.rootPath });
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// GET /api/browse-directory
-	if (url.pathname === "/api/browse-directory" && req.method === "GET") {
-		const rawPath = url.searchParams.get("path");
-		const dirPath = rawPath ? path.resolve(rawPath) : config.defaultCwd;
-
-		if (!fs.existsSync(dirPath)) {
-			json({ error: "Directory not found" }, 404);
-			return;
-		}
-
-		try {
-			const stat = fs.statSync(dirPath);
-			if (!stat.isDirectory()) {
-				json({ error: "Path is not a directory" }, 400);
-				return;
-			}
-		} catch {
-			json({ error: "Cannot access path" }, 400);
-			return;
-		}
-
-		const entries: Array<{ name: string; path: string }> = [];
-		try {
-			const items = fs.readdirSync(dirPath);
-			for (const item of items) {
-				// Skip hidden directories and node_modules
-				if (item.startsWith(".") || item === "node_modules") continue;
-				const fullPath = path.join(dirPath, item);
-				try {
-					const stat = fs.lstatSync(fullPath);
-					if (stat.isDirectory() && !stat.isSymbolicLink()) {
-						entries.push({ name: item, path: fullPath });
-					}
-				} catch {
-					// Skip entries we can't stat
-				}
-			}
-		} catch {
-			json({ error: "Cannot read directory" }, 500);
-			return;
-		}
-
-		entries.sort((a, b) => a.name.localeCompare(b.name));
-
-		const parsed = path.parse(dirPath);
-		const parent = parsed.root === dirPath ? null : path.dirname(dirPath);
-
-		json({ current: dirPath, parent, entries });
-		return;
-	}
-
 	// ── Project CRUD ──────────────────────────────────────────────────
-
-	// GET /api/projects
-	if (url.pathname === "/api/projects" && req.method === "GET") {
-		// Filter out hidden projects (e.g. the synthetic "system" project) so
-		// they never appear in the client's state.projects.
-		json(projectRegistry.list().filter(p => !p.hidden));
-		return;
-	}
 
 	// POST /api/projects
 	if (url.pathname === "/api/projects" && req.method === "POST") {
@@ -1460,84 +1292,6 @@ async function handleApiRoute(
 				});
 			}
 			json(project, 201);
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// GET /api/projects/:id
-	const projectGetMatch = url.pathname.match(/^\/api\/projects\/([^/]+)$/);
-	if (projectGetMatch && req.method === "GET") {
-		const project = projectRegistry.get(projectGetMatch[1]);
-		if (!project) { json({ error: "Project not found" }, 404); return; }
-		json(project);
-		return;
-	}
-
-	// PUT /api/projects/:id
-	if (projectGetMatch && req.method === "PUT") {
-		const body = await readBody(req);
-		const updates: { name?: string; color?: string; rootPath?: string; palette?: string; colorLight?: string; colorDark?: string } = {};
-		if (typeof body?.name === "string") updates.name = body.name;
-		if (typeof body?.color === "string") updates.color = body.color;
-		if (typeof body?.rootPath === "string") updates.rootPath = body.rootPath;
-		if (typeof body?.palette === "string" || body?.palette === null || body?.palette === "") updates.palette = body.palette ?? "";
-		if (typeof body?.colorLight === "string") updates.colorLight = body.colorLight;
-		if (typeof body?.colorDark === "string") updates.colorDark = body.colorDark;
-		try {
-			const updated = projectRegistry.update(projectGetMatch[1], updates);
-			json(updated);
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// DELETE /api/projects/:id
-	if (projectGetMatch && req.method === "DELETE") {
-		const projectId = projectGetMatch[1];
-		const project = projectRegistry.get(projectId);
-		// Test-only bypass: BOBBIT_E2E=1 + ?force=1 allows deleting the last
-		// project so GR-09 (zero-project first-run UX) can exercise the empty state.
-		const forceDelete = process.env.BOBBIT_E2E === "1" && url.searchParams.get("force") === "1";
-		// The synthetic "system" project (hidden) doesn't count toward the
-		// "at least one real project" guard.
-		const visibleCount = projectRegistry.list().filter(p => !p.hidden).length;
-		if (project && !project.hidden && visibleCount === 1 && !forceDelete) {
-			json({ error: "Cannot delete the last remaining project — add another project first" }, 400);
-			return;
-		}
-		try {
-			// Drain the project's worktree pool before removing
-			await sessionManager.removeWorktreePool(projectId);
-			// Terminate all live sessions belonging to the removed project
-			const liveSessions = sessionManager.listSessions().filter(s => s.projectId === projectId);
-			for (const s of liveSessions) {
-				try { await sessionManager.terminateSession(s.id); } catch {}
-			}
-			projectContextManager.remove(projectId);
-			if (project?.provisional) {
-				projectRegistry.removeProvisional(projectId);
-			} else {
-				projectRegistry.remove(projectId);
-			}
-			json({ ok: true });
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// POST /api/projects/:id/promote
-	const projectPromoteMatch = url.pathname.match(/^\/api\/projects\/([^/]+)\/promote$/);
-	if (projectPromoteMatch && req.method === "POST") {
-		const projectId = projectPromoteMatch[1];
-		try {
-			const body = await readBody(req);
-			const name = typeof body?.name === "string" ? body.name : undefined;
-			const promoted = projectRegistry.promote(projectId, { name });
-			json(promoted);
 		} catch (err: any) {
 			jsonError(400, err);
 		}
@@ -1814,15 +1568,6 @@ async function handleApiRoute(
 			json({ ok: true });
 			return;
 		}
-	}
-
-	// GET /api/projects/:id/qa-testing-config
-	const evConfigMatch = url.pathname.match(/^\/api\/projects\/([^/]+)\/qa-testing-config$/);
-	if (evConfigMatch && req.method === "GET") {
-		const ctx = projectContextManager.getOrCreate(evConfigMatch[1]);
-		if (!ctx) { json({ error: "Project not found" }, 404); return; }
-		json({ configured: ctx.projectConfigStore.isQaConfiguredOnAnyComponent() });
-		return;
 	}
 
 	// GET /api/search
